+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/exposition/ExpoLocal.java
@@ -43,7 +43,7 @@ public interface ExpoLocal {
     
     public Conducteur getConducteur(Long idConducteur);
     
-    public Duree creerDuree(int duree, Station station1, Station station2);
+    public Duree creerDuree(Long duree, Station station1, Station station2);
     
     public Duree getDuree(Long idDuree);
