+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/facades/DureeFacadeLocal.java
@@ -31,7 +31,7 @@ public interface DureeFacadeLocal {
 
     int count();
     
-    public Duree creerDuree(int duree, Station station1, Station station2);
+    public Duree creerDuree(Long duree, Station station1, Station station2);
     
     public Duree getDuree(Long idDuree);
