+++ Spacelib/Spacelib-ejb/src/main/java/fr/miage/m1/metier/GestionDuree.java
@@ -22,7 +22,7 @@ public class GestionDuree implements GestionDureeLocal {
     private DureeFacadeLocal dureeFacade;
 
     @Override
-    public Duree creerDuree(int duree, Station station1, Station station2) {
+    public Duree creerDuree(Long duree, Station station1, Station station2) {
         return this.dureeFacade.creerDuree(duree, station1, station2);
     }
